实现回复功能-异常处理

+++ src/main/java/com/example/helloworld/demo/Controller/QuestionController.java
@@ -1,8 +1,6 @@
 package com.example.helloworld.demo.Controller;
 
-import com.example.helloworld.demo.Model.Question;
 import com.example.helloworld.demo.dto.QuestionDTO;
-import com.example.helloworld.demo.mapper.QuestionMapper;
 import com.example.helloworld.demo.service.QuestionService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -17,7 +15,7 @@ public class QuestionController {
     private QuestionService questionService;
 
     @GetMapping("/question/{questionId}")
-    public String question(@PathVariable(name = "questionId") Integer questionId,
+    public String question(@PathVariable(name = "questionId") Long questionId,
                            Model model){
         QuestionDTO questionDTO = questionService.GetQuestionById(questionId);
         questionService.incViewCount(questionDTO.getId());
